extracted warm start implementation to a saperate class
